darwin-ops.c: Finished the 48px ico file generation code and stayed in CoreFoundation.framework and ImageIO.Framework. darwin-ops.h: Added ICO file structs. main.c: Minor code cleanups.

+++ LLMNRd/llmnrd.h
@@ -33,7 +33,7 @@ typedef struct {
     io_object_t           notification;
     CFStringRef           deviceName;
     uint8_t               hwAddress [ kIOEthernetAddressSize ];
-    uint16_t              ifType;        // The generic kernel interface type
+    uint32_t              ifType;        // The generic kernel interface type
                                          // (csma/cd applies to ethernet/firware
                                          // and wireless, although they are
                                          // different and wireless is CSMA/CA
